[refactor] response 전송 시 hashmap 활용, pk long타입으로 변경, service interface 추가

See merge request s04-ai-image-sub3/s04p23a202!11

+++ backend/fish-server/src/main/java/com/a202/fishserver/domain/fish/FishRepository.java
@@ -2,5 +2,5 @@ package com.a202.fishserver.domain.fish;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FishRepository extends JpaRepository<Fish, Integer> {
+public interface FishRepository extends JpaRepository<Fish, Long> {
 }
